move prim/mat/meta to prim/scal/meta
