--- stan/math/prim/mat/meta/value_type.hpp
@@ -1,24 +1,0 @@
-#ifndef STAN_MATH_PRIM_MAT_META_VALUE_TYPE_HPP
-#define STAN_MATH_PRIM_MAT_META_VALUE_TYPE_HPP
-
-#include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/mat/meta/is_eigen.hpp>
-#include <stan/math/prim/scal/meta/value_type.hpp>
-#include <type_traits>
-
-namespace stan {
-
-/** \ingroup type_trait
- * Template metaprogram defining the type of values stored in an
- * Eigen matrix, vector, or row vector.
- *
- * @tparam T type of matrix.
- */
-template <typename T>
-struct value_type<T, std::enable_if_t<is_eigen<T>::value>> {
-  using type = typename std::decay_t<T>::Scalar;
-};
-
-}  // namespace stan
-
-#endif
