+++ stan/math/prim/scal/meta/append_return_type.hpp
