+++ stan/math/prim/scal/meta/broadcast_array.hpp
@@ -1,6 +1,8 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_BROADCAST_ARRAY_HPP
 #define STAN_MATH_PRIM_SCAL_META_BROADCAST_ARRAY_HPP
 
+#include <stan/math/prim/scal/meta/broadcast_array.hpp>
+#include <stan/math/prim/mat/fun/Eigen.hpp>
 #include <stdexcept>
 
 namespace stan {
@@ -43,8 +45,41 @@ class empty_broadcast_array {
   template <typename Y>
   void operator=(const Y& /*A*/);
 };
+
+template <typename ViewElt, typename OpElt, int R, int C>
+class empty_broadcast_array<ViewElt, Eigen::Matrix<OpElt, R, C> > {
+ public:
+  empty_broadcast_array() {}
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  ViewElt& operator[](int /*i*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  ViewElt& operator()(int /*i*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  void operator=(const Eigen::Matrix<ViewElt, R, C>& /*A*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  void operator+=(Eigen::Matrix<ViewElt, R, C> /*A*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  void operator-=(Eigen::Matrix<ViewElt, R, C> /*A*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  Eigen::Matrix<ViewElt, 1, C>& row(int /*i*/);
+  /** \ingroup type_trait
+   * Not implemented so cannot be called.
+   */
+  Eigen::Matrix<ViewElt, R, 1>& col(int /*i*/);
+};
 }  // namespace internal
 }  // namespace math
 }  // namespace stan
-
 #endif
