+++ stan/math/prim/scal/meta/value_type.hpp
@@ -1,6 +1,8 @@
 #ifndef STAN_MATH_PRIM_SCAL_META_VALUE_TYPE_HPP
 #define STAN_MATH_PRIM_SCAL_META_VALUE_TYPE_HPP
 
+#include <stan/math/prim/mat/fun/Eigen.hpp>
+#include <stan/math/prim/scal/meta/is_eigen.hpp>
 #include <stan/math/prim/scal/meta/is_vector.hpp>
 #include <stan/math/prim/scal/meta/value_type.hpp>
 #include <type_traits>
@@ -45,5 +47,17 @@ struct value_type<T, std::enable_if_t<is_std_vector<T>::value>> {
   using type = typename std::decay_t<T>::value_type;
 };
 
+/** \ingroup type_trait
+ * Template metaprogram defining the type of values stored in an
+ * Eigen matrix, vector, or row vector.
+ *
+ * @tparam T type of matrix.
+ */
+template <typename T>
+struct value_type<T, std::enable_if_t<is_eigen<T>::value>> {
+  using type = typename std::decay_t<T>::Scalar;
+};
+
 }  // namespace stan
+
 #endif
